Create README.md
This little project consists of a Node.js rest api with no backend framework. it's a simple application with basic crud functionalities. I made the app to challenge myself to not use express or any external database (MongoDB, MySQL)). However, I did use React in the frontend for the UI.